requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
colorama==0.4.6
cryptography==42.0.5
cycler==0.12.1
dill==0.3.4
Django==3.2.2
django-webpush==0.3.5
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2024.2.0
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
gym==0.21.0
http-ece==1.2.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
kiwisolver==1.4.5
kombu==5.3.5
lightning-utilities==0.10.1
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.5.2
multidict==6.0.5
numpy==1.22.3
oauthlib==3.2.2
packaging==23.2
pandas==1.4.2
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psycopg2==2.9.9
py-vapid==1.9.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyDeprecate==0.3.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==1.6.3
pytz==2024.1
pywebpush==1.9.4
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stable-baselines3==1.5.0
tabulate==0.8.9
tensorboard==2.15.2
tensorboard-data-server==0.7.2
torch==1.11.0
torchmetrics==1.3.0.post0
tqdm==4.64.0
typing-extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
